fix typo and add up to step 18
