+++ packages/modules/pendle/src/actions/on-chain/subactions/setPendleRouterV4Facets.ts
@@ -2,41 +2,43 @@ import { Address, Hex } from 'viem'
 
 import { InfinitWallet, SubAction, SubActionExecuteResponse } from '@infinit-xyz/core'
 
-import { SetFacetForSelectorsTxBuilder } from '@actions/on-chain/subactions/txBuilders/PendleRouterStatic/setFacetForSelectors'
+import { SetSelectorToFacetsTxBuilder } from '@actions/on-chain/subactions/txBuilders/PendleRouter/setSelectorToFacets'
 
 import { PendleV3Registry } from '@/src/type'
 
-export type SetPendleRouterStaticFacetsParams = {
-  pendleRouterStatic: Address
-  actionStorageStatic: Address
-  actionInfoStatic: Address
-  actionMarketAuxStatic: Address
-  actionMarketCoreStatic: Address
-  actionMintRedeemStatic: Address
-  actionVePendleStatic: Address
+export type SetPendleRouterV4FacetsParams = {
+  pendleRouterV4: Address
+  actionStorageV4: Address
+  actionAddRemoveLiqV3: Address
+  actionCallbackV3: Address
+  actionMiscV3: Address
+  actionSimple: Address
+  actionSwapPTV3: Address
+  actionSwapYTV3: Address
 }
 
-export type SetPendleRouterStaticFacetsMsg = {}
+export type SetPendleRouterV4FacetsMsg = {}
 
-export class SetPendleRouterStaticFacets1SubAction extends SubAction<
-  SetPendleRouterStaticFacetsParams,
+export class SetPendleRouterV4FacetsSubAction extends SubAction<
+  SetPendleRouterV4FacetsParams,
   PendleV3Registry,
-  SetPendleRouterStaticFacetsMsg
+  SetPendleRouterV4FacetsMsg
 > {
-  constructor(client: InfinitWallet, params: SetPendleRouterStaticFacetsParams) {
-    super(SetPendleRouterStaticFacets1SubAction.name, client, params)
+  constructor(client: InfinitWallet, params: SetPendleRouterV4FacetsParams) {
+    super(SetPendleRouterV4FacetsSubAction.name, client, params)
   }
 
   protected setTxBuilders(): void {
     this.txBuilders.push(
-      new SetFacetForSelectorsTxBuilder(this.client, {
-        pendleRouterStatic: this.params.pendleRouterStatic,
-        actionStorageStatic: this.params.pendleRouterStatic,
-        actionInfoStatic: this.params.pendleRouterStatic,
-        actionMarketAuxStatic: this.params.pendleRouterStatic,
-        actionMarketCoreStatic: this.params.pendleRouterStatic,
-        actionMintRedeemStatic: this.params.pendleRouterStatic,
-        actionVePendleStatic: this.params.pendleRouterStatic,
+      new SetSelectorToFacetsTxBuilder(this.client, {
+        pendleRouter: this.params.pendleRouterV4,
+        actionStorageV4: this.params.actionStorageV4,
+        actionAddRemoveLiqV3: this.params.actionAddRemoveLiqV3,
+        actionCallbackV3: this.params.actionCallbackV3,
+        actionMiscV3: this.params.actionMiscV3,
+        actionSimple: this.params.actionSimple,
+        actionSwapPTV3: this.params.actionSwapPTV3,
+        actionSwapYTV3: this.params.actionSwapYTV3,
       }),
     )
   }
@@ -44,7 +46,7 @@ export class SetPendleRouterStaticFacets1SubAction extends SubAction<
   public async updateRegistryAndMessage(
     registry: PendleV3Registry,
     _txHashes: Hex[],
-  ): Promise<SubActionExecuteResponse<PendleV3Registry, SetPendleRouterStaticFacetsMsg>> {
+  ): Promise<SubActionExecuteResponse<PendleV3Registry, SetPendleRouterV4FacetsMsg>> {
     return { newRegistry: registry, newMessage: {} }
   }
 }
